Fixed tickets issue

+++ frontend/lib/routes/session/AddSessionForm.dart
@@ -64,7 +64,7 @@ class _AddSessionFormState extends State<AddSessionForm> {
       var place = _placeController.text;
       //var speaker = _speakerController.text;
       //var company = _companyController.text;
-      var maxTickets = _currentTicketsValue as int;
+      var maxTickets = _currentTicketsValue;
       var videoURL = _videoURLController.text;
 
       print("Max tickets:");
@@ -76,7 +76,7 @@ class _AddSessionFormState extends State<AddSessionForm> {
       //     : new SessionTickets(max: 0, start: null, end: null);
 
       var sessionTickets = new SessionTickets(
-          max: maxTickets, start: _beginTicket, end: _endTicket);
+          max: maxTickets as int, start: _beginTicket, end: _endTicket);
 
       ScaffoldMessenger.of(context).showSnackBar(
         const SnackBar(content: Text('Uploading')),
